backend/requirements.txt
gunicorn==20.1.0
requests==2.26.0
Django==3.2.3
django-filter==23.2
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
Pillow==9.0.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
djoser==2.1.0
webcolors==1.11.1
PyYAML==6.0
